+++ src/feature/CardRegister/hooks/useExpiryDate.ts
@@ -1,5 +1,5 @@
 import {useState} from 'react';
-import {createFlags, computeNextTouched, computeNextErrorInfo} from './fieldState';
+import {createFlags, computeNextTouched, computeNextErrorInfo} from './fieldErrorUtils';
 
 const ERROR_MSG = '2자리를 입력해 주세요';
 const INVALID_MONTH_MSG = '01~12 사이의 월을 입력해 주세요';
